requirements.txt
numpy
opencv-python-headless==4.13.0.92
rpi-hardware-pwm==0.3.1
mavsdk==2.8.2
picamera2==0.3.27
PyYAML==6.0.2
